Add files via upload
Attempt #1:- Tried fixing the glichy beasts sprites by assigning a specified value for the minimum separation between the player and the enemy(The value is currently set to a very large value for the purpose of testing.) If the enemy sprite enters the minimum separation radius, it will try to move out of the radius and will only attack outside of the radius. 
Status:- Failure.